Create TradeStocks
This is where we combine our stock data that we scraped, plus the model results, to determine if there is a good opportunity to buy/sell any of our stocks. This is done by looking at the predicted High and Low price for the next hour of each stock, and if the stock's predicted high AND low prices are both higher than the latest high and low prices, we buy the stock. On the contrary, if we own a current stock and the predicted high AND low prices are both lower than the latest high and low prices, we sell the stock. I also have limit prices and stop prices added in as well for any owned stocks. 

On the first run, you will need to keep the "create_order(...)" parameter within the "Order_bulk()" definition commented out, and uncomment the "create_first_order(...)" parameter. After the first run, you can uncomment the "create_order(...)" parameter and comment out the "create_first_order(...)". 

I use Alpaca Trading platform, which allows for both live trading and paper trading, and the URL will need to be updated accordingly. 


take our top 20 stocks who are run through our model, and determine if our model determines there is a good opportunity to purchase